Update set core version script to set wp config.
